Written by: Simon Bertrang <simon@openbsd.org>
Major update to version 8.3.1 - be sure to dump your databases before
you apply this and restore afterwards!

Additionally, implicit typecasts are history and not supported anymore.
Versions prior to 8.3 had the feature (some say bug) that functions,
expecting an argument to be of a certain type, have casted a variable of
any other type to the expected type, if possible.
This has changed now.  Tests surfaced rare occurrences of regressions,
which were then fixed in about ten minutes - and that code was not even
in the ports tree;  no issues found there.

A few more things have changed, namely tsearch2 went from contrib to the
core and native uuid type support was added;  for details read the
release announcement at
http://www.postgresql.org/docs/8.3/static/release-8-3.html.

tests & ok mbalmer@ (maintainer)