Written by: Okan Demirmen <okan@openbsd.org>
import Geo::METAR

METAR reports are available on-line, thanks to the National Weather
Service.  Since reading the METAR format isn't easy for non-pilots,
these reports are relatively useles to the common man who just wants a
quick glace at the weather.  This module tries to parse the METAR
reports so the data can be used to create readable weather reports
and/or process the data in applications.

feedback and ok merdely@