Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gnome-scan-0.6

The GNOME Scan project aims to provide scanning features integrated
throughout the GNOME desktop, similar to printing.  The GNOME Scan
project consists of a C/GObject library upon which are built standalone
applications and plugins for various programs.
The application and plugins are called flegita.


Note:
gnome-scan is still in its early stage. It is not a drop-in
replacement for xsane or alikes (yet); it more suited for for mass
acquisition and ADF scanners.
There's an issue with image enhancement which is being worked on
upstream and the gsane backend may not give good results on all
scanners.

ok jasper@