Written by: Stuart Henderson <sthen@openbsd.org>
Use a floating window for xonix, otherwise dwm forces the window
size and xonix segfaults. (Apologies to anyone who was using this
dwm behaviour as a way to keep them from getting addicted...)

ok Jim Razmus (maintainer) with a suggestion to use xprop to find
the Class rather than the window title (as my original diff had).
Also thanks to deanna@ for suggestions with the segfault.