Written by: Bernd Ahlers <bernd@openbsd.org>
Unbreak one half of evolution by not installing the evolution-data-server
binary suid root.

This caused lots of problems with the calendar and other stuff because the
evolution-data-server didn't run correctly.

Well, evolution-data-server is currently broken because of the libsoup
major update. It needs an update to a more recent version.