Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import py-gdata-1.0.13

The Google data Python client library makes it easy to access data
through the Google data APIs. This library provides data model and
service modules for the the following Google data services:
 - Google Calendar data API
 - Google Contacts data API
 - Google Spreadsheets data API
 - Google Document List data APIs
 - Google Base data API
 - Google Apps Provisioning API
 - Picasa Web Albums Data API
 - Google Code Search Data API
 - core Google data API functionality

The core Google data code provides sufficient functionality to use this
library with any Google data API (even if a module hasn't been written
for it yet). For example, this client can be used with the Blogger API,
and the YouTube API. This library may also be used with any Atom
Publishing Protocol service.