Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import totem-2.22.2... under the correct dir this time

Totem is the official movie player of the GNOME desktop environment
based on xine-lib or GStreamer. It features a playlist, a full-screen
mode, seek and volume controls, as well as keyboard navigation.

Available flavor:
        xine - Use xine backend instead of GStreamer

loosely based on an older port from jasper@
ok jasper@