Written by: Marc Espie <espie@openbsd.org>
yeah, the gstreamer people finally made it possible to use the system
ffmpeg.

Note that their fuzzy warnings don't apply to us. This is so stupid.
Between the ffmpeg people, who only care for performance to the exclusion
of any sense (like, good luck making it work elsewhere), and the gstreamer
people, who don't care that stuff won't work elsewhere.