Written by: Marc Balmer <mbalmer@openbsd.org>
mimetex is a CGI program to convert LaTeX math formulae into gif images
without the need to install LaTeX in httpds chroot.

From Mischa Diehm <md@msys.ch>, maintainer (a new porter, so give him the
drill he needs...)

Thanks to merdely to host the distfile and rerolling it as a gzipped tarball.