Written by: Steven Mestdagh <steven@openbsd.org>
import tmux 0.2

tmux is a "terminal multiplexer", it enables a number of terminals (or 
windows) to be accessed and controlled from a single terminal. tmux is 
intended to be a simple, modern, BSD-licensed alternative to programs 
such as GNU screen.

port made by brynet at gmail, and Nicholas Marriott (maintainer)