Written by: Stuart Henderson <sthen@openbsd.org>
SECURITY update to fix CVE-2008-0960, http://secunia.com/advisories/30574/

A vulnerability has been reported in Net-SNMP, which can be exploited
by malicious people to spoof authenticated SNMPv3 packets.

The authentication code reads the length to be checked from sender
input, this allows the sender to supply single byte HMAC code and
have a 1 in 256 chance of matching the correct HMAC and authenticating,
as only the first byte will be checked. The sender would need to
know a valid username.

ok rui (maintainer)