Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of jruby-1.1.2.

JRuby is an 100% pure-Java implementation of the Ruby programming language.

JRuby is tightly integrated with Java to allow the embedding of the
interpreter into any Java application with full two-way access between the
Java and the Ruby code.

help & testing kurt@, ian@ and sturm@

ok kurt@