Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import Record My Desktop (and gtk-frontend)

Recordmydesktop produces a file(default out.ogg) that contains a video
and audio recording of a desktop session. The default behavior of
recording is to mark areas that have changed(through libxdamage) and
update the frame.

with patch from bernd@ to make sound recording work.