Written by: Peter Hessler <phessler@openbsd.org>
Aleph One is a game engine currently in development by the Open Source 
community. It is based on the source code of Marathon 2: Durandal, a 
game created by Bungie Software in 1995. Since Bungie released the 
source code, Aleph One has matured to include a variety of new features 
and improvements. 

ok merdely