Written by: Will Maier <wcmaier@openbsd.org>
DRIconf is a configuration applet for the Direct Rendering
Infrastructure. It allows customizing performance and visual quality
settings of OpenGL drivers on a per-driver, per-screen and/or
per-application level.

The settings are stored in system wide and per-user XML configuration
files, which are parsed by the OpenGL drivers on startup.

Originally by oga@; tested by oga@ and guenther@. Python jiggery and
other cleanups by moi.

double ok fgsch@, ribbit landry@