Written by: Marc Espie <espie@openbsd.org>
new update-or-install target needs to have dependent subpackages built as
well.

Incidentally, this is also an issue with update: in some cases, the update
target wouldn't work (specifically, when dependencies had changed).
Fix it as well.

There might be a need to revisit the DEPENDS_TARGET stuff later, this is
a bit of a kludge...