Written by: Stuart Henderson <sthen@openbsd.org>
configure tries to detect isnan() as a function in libm to work out
whether it should include it. isnan is now a macro so this fails.

upstream configure.ac _should_ be rewritten to use pkg-config but,
for the sake of expediency, workaround by adding to LDFLAGS for
now, as was already done to fix the libvorbis test. fixes the build.

ok naddy@