Written by: Stuart Henderson <sthen@openbsd.org>
import cunit:

CUnit is a lightweight system for writing, administering, and running
unit tests in C.  It provides C programmers a basic testing
functionality with a flexible variety of user interfaces.  It uses a
simple framework for building test structures, and provides a rich set
of assertions for testing common data types.   In addition, several
different interfaces are provided for running tests and reporting
results.

from maintainer Markus Bergkvist, thanks!