Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import pyzor 0.4.0

Pyzor is a collaborative, networked system to detect and block spam
using identifying digests of messages.

Pyzor initially started out to be merely a Python implementation of
Razor, but due to the protocol and the fact that Razor's server is not
Free Software, I decided to Pyzor should be based upon a new, lighter,
more language- independent protocol.

requested by and ok dlg@