Written by: Brad Smith <brad@openbsd.org>
Fix FFmpeg's configure script so it can retreive and understand the
hardware architecture instead of the machine port. Also add support
for detecting the presence of AltiVec at run-time using the
machdep.altivec sysctl.

At the moment AltiVec is disabled but this at least allows the
configure script to build in the support if --disable-altivec is
removed. Facilitates further testing and debugging.

ok kili@ jakemsr@