Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import libneural 1.0.3

libneural is an extremely tiny library for creating a software three
layer backpropagation neural network. This is useful for a very wide
variety of pattern recognition and classification problems. It is
written in C++, based on the standard algorithm with NO fancy features
like bias terms or momentum.