Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import tesseract-2.03

The Tesseract OCR engine was one of the top 3 engines in the 1995 UNLV
Accuracy test. Between 1995 and 2006 it had little work done on it, but
it is probably one of the most accurate open source OCR engines
available. The source code will read a binary, grey or color image and
output text. A tiff reader is built in that will read uncompressed TIFF
images, or libtiff can be added to read compressed images.


reworked from scratch from an original submission by
jrrs at ice-nine dot org a while ago on ports@; he's been non-responsive
so far so I'm taking MAINTAINERship.