Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import powerdns 2.9.21.1

pdnsd is a proxy dns server with permanent caching (the cache
contents are written to hard disk on exit) that is designed to cope
with unreachable or down dns servers (for example in dial-in
networking).

original port by dlg@ and robert@, banged into shape by me.
not hooked to the builds as this needs further testing/tweaking