Written by: Robert Nagy <robert@openbsd.org>
import of mono-2.0:

Mono is an open source implementation of .NET Development Framework.
Its objective is to enable UNIX developers to build and deploy cross-platform
.NET Applications. The project implements various technologies developed by
Microsoft that have now been submitted to the ECMA for standardization.

Mono provides the necessary software to develop and run .NET client and
server applications on BSD, Linux, Solaris, Mac OS X, Windows and Unix.

from alek@, modifications by me and some amd64 patches from
Giovanni Bechis

mono is not linked to the build yet, it's getting imported so we can
work on it in tree;