Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import Thinking Forth

Thinking Forth is a book about the philosophy of problem solving and
programming style, applied to the unique programming language Forth.
Published first in 1984, it could be among the timeless classics of
computer books, such as Fred Brooks' The Mythical Man-Month and Donald
Knuth's The Art of Computer Programming.