Written by: Paul Irofti <pirofti@openbsd.org>
Initial import of ruby-sexp_processor.

This is the sexp_processor module for parsetree.

ParseTree is a C extension (using RubyInline) that extracts the parse tree for
an entire class or a specific method and returns it as a s-expression (aka sexp)
using ruby's arrays, strings, symbols, and integers.

Okay bernd@.