Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import xdg-utils-1.0.2

Xdg-utils is a set of command line tools that assist applications with a
variety of desktop integration tasks.  About half of the tools focus on
tasks commonly required during the installation of a desktop application
and the other half focuses on integration with the desktop environment
while the application is running.

based on an unfinished port from landry@
ok landry@