Written by: Brad Smith <brad@openbsd.org>
Upgrade to MailScanner 4.72.5-1.

Feedback from sthen@ including adding a conflict marker as a
kludge to ensure that if a user wants to use AV support that
the correct version is installed (0.94 or newer) since it is
an optional dependency. Great job by the Clam crew for making
this such a mess. The current version of MailScanner is broken
if using 0.94 since the commandline opts were changed.

ok todd@ sthen@