Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import gnome-settings-daemon 2.24.0 (not linked to the builds yet)

GNOME settings deamon to control settings and preferences throughout the
desktop.

this used to be part of gnome-control-center, but it was split of.
it depends on some stuff not committed yet (e.g. gnome-desktop update),
but having it in tree allows the gtk-frogs to test it more easily.