Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import matchbox-nest 0.3

Matchbox-nest is a graphical wrapper around xnest. You can make Xnest
look like a particular device's display and set up buttons on that
device.

if someone wants to provide a frontend in the style of a Zaurus,
please drop me a mail.

test building and "just commit it" kili@, thanks!