Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import bluez-hcidump 1.42

HCIdump reads raw HCI data coming from and going to a Bluetooth device
and prints to screen commands, events and data in a human-readable form.
Optionally, the dump can be written to a file rather than parsed, and
the dump file can be parsed in a subsequent moment.

from jcs@, with some tweaks by me
requested by uwe@