Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
so long and thanks for all the...porn?

remove pornview as it's and old, unmaintained port with code that's bad and
doesn't play nice with gtk2. there are enough ports that do the same (but
don't have the nice name)... gqview, ristretto, gthumb, gpicview, etc,etc)

as discussed with tedu@, landry@ and ajacoutot@
"at this point, I won't argue for keeping it alive." tedu@
ajacoutot@ had a long story about dragons and eating virgins, but basicly "ok"