Written by: Landry Breuil <landry@openbsd.org>
Update to gtk2mm 2.14.3. As it is still broken and waiting for a
libstdc++ fix to be commited, face reality and mark it BROKEN.
Hopefully libstdc++ will be fixed real soon.
Based on a diff flying around since p2k8.

discussed at length with, and agreed by ajacoutot@ espie@ jasper@