Written by: Matthias Kilian <kili@openbsd.org>
Build, install and use libgs in all flavors on shared archs. Note:
there is currently no libgs (not even a libgs.a) on static archs.

Install only *one* binary (bin/gs), even in the gtk flavor; the gsc
binary in the gtk flavor was identically to gsx, with the exception
that the display driver (which is the sole purpose of the gtk flavor)
didn't work in gsc.

ok ajacoutot@