Written by: Pierre-Yves Ritschard <pyr@openbsd.org>
Import scrotwm.
``Scrotwm is a small dynamic tiling window manager for X11. It tries to
stay out of the way so that valuable screen real estate can be used for
much more important stuff. It has sane defaults and does not require one
to learn a language to do any configuration. It was written by hackers
for hackers and it strives to be small, compact and fast. ''

Development is ongoing, but I'll make (or you can tell me to make)
snapshots every time it is necessary.

ok ajacoutot@ and other members of the french family.