Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of ruby-merb-slices-1.0.8.1.

Merb-Slices is a Merb plugin for using and creating application 'slices' which
help you modularize your application. Usually these are reusable extractions
from your main app. In effect, a Slice is just like a regular Merb MVC
application, both in functionality as well as in structure.