Written by: Stuart Henderson <sthen@openbsd.org>
somewhere in the update between 4.20 and 4.28 upstream changed this to
install an identical copy of the 6x12 font named "bold" (i.e. to prevent
a synthetic bold font being generated). everyone I talked to hates this,
and would like it if 6x12 bold was distinct from 6x12 normal, even if
it's a little blocky, so change ALL_TARGET from pcf to n12, and regen
the PLIST, this reinstates the expected behaviour.