Written by: Stuart Henderson <sthen@openbsd.org>
the path in gui.c is used to find both wish (installed by tk which is a
different port, so, LOCALBASE), _and_ x11vnc (installed by the current port,
so PREFIX). fixit accordingly, for correctness and to cater for whoever
might like the pain of installing different ports with different prefixes.

discussed with/ok ajacoutot