Written by: Stefan Sperling <stsp@openbsd.org>
Don't install man pages for dvdrip-progress and dvdrip-splitpipe,
because they are in fact empty files.
The other man pages installed by the port are generated from
perl sources and are proper man pages. But these two programs
are written in C, so man page generation fails for them.
ok ajacoutot@