Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of apache-couchdb-0.8.1.

CouchDB is a distributed, document-oriented, Non-Relational DataBase
Management Server (NRDBMS).

It is accessible via a RESTful JSON API, and provides incremental
replication with bi-directional conflict detection and management.
Data is stored with a schema-free, flat address space, and may be
queried via a table-oriented reporting engine that uses Javascript
as a query language.

Joint work of maintainer Benoit Chesneau, fgsch@ and myself.